+++ CHANGELOG.md
@@ -52,8 +52,10 @@ Removes preceding `v` in version name to be like `0.3.2`.
     - Companion Card:
       - Chang the Ninth
       - Liben
-      - NRE
       - Paimon
+    - Item Card:
+      - NRE
+      - Parametric Transformer
   - Event Card:
     - Where Is the Unseen Razor?